add channel and thread events

+++ integration-tests/channels.ts
@@ -1,36 +1,58 @@
 import { Discord } from "botloader";
 import { assertExpected, runOnce, sendScriptCompletion } from "lib";
 
+const bot = Discord.getBotUser();
+
+script.on("CHANNEL_CREATE", async (channel) => {
+    if (channel.name === "bl-chtest-1") {
+        // give ourselves send messages
+        await Discord.editChannelPermission(channel.id,
+            Discord.PermissionOverwrite.member(bot.id, Discord.Permissions.SendMessages,
+                new Discord.Permissions())
+        );
+
+        // put this channel under a category
+        let edited = await Discord.editChannel(channel.id, {
+            parentId: "531120790318350337",
+        });
+
+        if ("permissionOverwrites" in edited) {
+            assertExpected(2, edited.permissionOverwrites.length);
+        } else {
+            throw new Error("No permission overwrites...");
+        }
+    }
+})
+
+script.on("CHANNEL_UPDATE", async (channel) => {
+    if (channel.name === "bl-chtest-1" && "parentId" in channel && channel.parentId === "531120790318350337") {
+        if ("permissionOverwrites" in channel) {
+            if (channel.permissionOverwrites.length === 2) {
+                await Discord.deleteChannelPermission(channel.id, "Member", bot.id);
+            } else if (channel.permissionOverwrites.length === 1) {
+                await Discord.deleteChannel(channel.id);
+            } else {
+                throw new Error("unknown number of overwrites!");
+            }
+        } else {
+            throw new Error("No permission overwrites...");
+        }
+    }
+})
+
+script.on("CHANNEL_DELETE", async (channel) => {
+    if (channel.name === "bl-chtest-1") {
+        sendScriptCompletion();
+    }
+})
+
 runOnce("channels.ts", async () => {
-    let bot = Discord.getBotUser();
     let channel = await Discord.createChannel({
-        name: "gaming",
+        name: "bl-chtest-1",
         topic: "We are gaming",
         permissionOverwrites: [
             // diasllow send messages for everyone
             Discord.PermissionOverwrite.everyone(new Discord.Permissions(), Discord.Permissions.SendMessages),
         ]
     });
-
-    // give ourselves send messages
-    await Discord.editChannelPermission(channel.id,
-        Discord.PermissionOverwrite.member(bot.id, Discord.Permissions.SendMessages,
-            new Discord.Permissions())
-    );
-
-    // put this channel under a category
-    let edited = await Discord.editChannel(channel.id, {
-        parentId: "531120790318350337",
-    });
-
-    if ("permissionOverwrites" in edited) {
-        assertExpected(2, edited.permissionOverwrites.length);
-    } else {
-        throw new Error("No permission overwrites...");
-    }
-
-    await Discord.deleteChannelPermission(channel.id, "Member", bot.id);
-    await Discord.deleteChannel(channel.id);
-
-    sendScriptCompletion();
 })
